+++ Plugins/NativeXr/Source/Constants.h
@@ -80,5 +80,5 @@ namespace Babylon
                 }
             }
         };
-    } // Plugins
+    }
 } // Babylon
