+++ Plugins/NativeXr/Source/NativeXrImpl.h
@@ -1,3 +1,5 @@
+#pragma once
+
 namespace Babylon
 {
     namespace Plugins
