+++ Plugins/NativeXr/Source/XRUtils.h
@@ -5,29 +5,6 @@
 
 namespace
 {
-    bgfx::TextureFormat::Enum XrTextureFormatToBgfxFormat(xr::TextureFormat format)
-    {
-        switch (format)
-        {
-            // Color Formats
-            // NOTE: Use linear formats even though XR requests sRGB to match what happens on the web.
-            //       WebGL shaders expect sRGB output while native shaders expect linear output.
-            case xr::TextureFormat::BGRA8_SRGB:
-                return bgfx::TextureFormat::BGRA8;
-            case xr::TextureFormat::RGBA8_SRGB:
-                return bgfx::TextureFormat::RGBA8;
-
-            // Depth Formats
-            case xr::TextureFormat::D24S8:
-                return bgfx::TextureFormat::D24S8;
-            case xr::TextureFormat::D16:
-                return bgfx::TextureFormat::D16;
-
-            default:
-                throw std::runtime_error{"Unsupported texture format"};
-        }
-    }
-
     // clang-format off
     constexpr std::array<float, 16> IDENTITY_MATRIX{
         1.f, 0.f, 0.f, 0.f,
