+++ Plugins/NativeXr/Source/XRWebGLBinding.h
@@ -31,5 +31,5 @@ namespace Babylon
             {
             }
         };
-    } // Plugins
+    }
 } // Babylon
